requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
lxml>=4.9.0
lxml_html_clean>=0.1.0

# YouTube
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.1

# PubMed
biopython>=1.83

# NLP & Topic Tagging
keybert>=0.8.3
scikit-learn>=1.3.0

# Language Detection
langdetect>=1.0.9

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0
instructor>=1.2.6
pydantic>=2.7.0
openai>=1.20.0
python-dotenv>=1.0.1
